Redirect to return URL even if auth fails
