+++ apps/antalmanac/src/backend/lib/helpers.ts
@@ -116,3 +116,17 @@ export async function queryHTTPS<PromiseReturnType>(
 
     return json as PromiseReturnType;
 }
+
+export const getCookiesFromHeader = (headers: Headers) => {
+    const cookieHeader = headers.get('cookie') ?? '';
+    const cookies = Object.fromEntries(
+        cookieHeader
+            .split('; ')
+            .filter((c) => c.includes('='))
+            .map((c) => {
+                const [key, ...v] = c.split('=');
+                return [key, v.join('=')];
+            })
+    );
+    return cookies;
+};
